# openapi/rc-daemon.json
{
    "components": {
        "parameters": {
            "ActiveWorkspaceHeader": {
                "description": "Active workspace identifier for browser-scoped daemon requests.",
                "in": "header",
                "name": "X-rc-Workspace-ID",
                "required": true,
                "schema": {
                    "type": "string"
                }
            },
...
            "Limit": {
                "description": "Maximum number of runs to return.",
                "in": "query",
                "name": "limit",
                "schema": {
                    "default": 100,
                    "maximum": 500,
                    "minimum": 1,
                    "type": "integer"
                }
            },
...
            "RunMode": {
                "description": "Optional run mode filter.",
                "in": "query",
                "name": "mode",
                "schema": {
                    "type": "string"
                }
            },
            "RunStatus": {
                "description": "Optional run status filter.",
                "in": "query",
                "name": "status",
                "schema": {
                    "type": "string"
                }
            },
...
            "WorkspaceFilterQuery": {
                "description": "Optional workspace filter for run inventory reads.",
                "in": "query",
                "name": "workspace",
                "schema": {
                    "type": "string"
                }
            },
...
        },
        "responses": {
...
            "Forbidden": {
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/TransportError"
                        }
                    }
                },
                "description": "Request was rejected by browser-facing transport security."
            },
            "InternalError": {
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/TransportError"
                        }
                    }
                },
                "description": "Unexpected daemon failure."
            },
            "InvalidRequest": {
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/TransportError"
                        }
                    }
                },
                "description": "Request body or parameters were malformed."
            },
            "NotFound": {
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/TransportError"
                        }
                    }
                },
                "description": "Requested resource was not found."
            },
            "StaleWorkspace": {
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/TransportError"
                        }
                    }
                },
                "description": "Active workspace context is missing or stale."
            },
            "ValidationError": {
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/TransportError"
                        }
                    }
                },
                "description": "Request validation failed."
            }
        },
        "schemas": {
...
            "AgentItem": {
                "properties": {
                    "description": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "scope": {
                        "type": "string"
                    },
                    "warnings": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array"
                    }
                },
                "required": ["name", "scope"],
                "type": "object"
            },
            "AgentListResponse": {
                "properties": {
                    "agents": {
                        "items": {
                            "$ref": "#/components/schemas/AgentItem"
                        },
                        "type": "array"
                    }
                },
                "required": ["agents"],
                "type": "object"
            },
...
            "ConfigDefaults": {
                "properties": {
                    "access_mode": {
                        "type": "string"
                    },
                    "add_dirs": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array"
                    },
                    "auto_commit": {
                        "type": "boolean"
                    },
                    "ide": {
                        "type": "string"
                    },
                    "max_retries": {
                        "type": "integer"
                    },
                    "model": {
                        "type": "string"
                    },
                    "output_format": {
                        "type": "string"
                    },
                    "reasoning_effort": {
                        "type": "string"
                    },
                    "retry_backoff_multiplier": {
                        "type": "number"
                    },
                    "tail_lines": {
                        "type": "integer"
                    },
                    "timeout": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "ConfigDocument": {
                "properties": {
                    "defaults": {
                        "$ref": "#/components/schemas/ConfigDefaults"
                    },
                    "exec": {
                        "$ref": "#/components/schemas/ConfigExec"
                    },
                    "fetch_reviews": {
                        "$ref": "#/components/schemas/ConfigFetchReviews"
                    },
                    "fix_reviews": {
                        "$ref": "#/components/schemas/ConfigFixReviews"
                    },
                    "runs": {
                        "$ref": "#/components/schemas/ConfigRuns"
                    },
                    "sound": {
                        "$ref": "#/components/schemas/ConfigSound"
                    },
                    "tasks": {
                        "$ref": "#/components/schemas/ConfigTasks"
                    },
                    "watch_reviews": {
                        "$ref": "#/components/schemas/ConfigWatchReviews"
                    }
                },
                "type": "object"
            },
            "ConfigDocumentResponse": {
                "properties": {
                    "config": {
                        "$ref": "#/components/schemas/ConfigDocument"
                    }
                },
                "required": ["config"],
                "type": "object"
            },
            "ConfigExec": {
                "properties": {
                    "access_mode": {
                        "type": "string"
                    },
                    "add_dirs": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array"
                    },
                    "auto_commit": {
                        "type": "boolean"
                    },
                    "ide": {
                        "type": "string"
                    },
                    "max_retries": {
                        "type": "integer"
                    },
                    "model": {
                        "type": "string"
                    },
                    "output_format": {
                        "type": "string"
                    },
                    "persist": {
                        "type": "boolean"
                    },
                    "reasoning_effort": {
                        "type": "string"
                    },
                    "retry_backoff_multiplier": {
                        "type": "number"
                    },
                    "tail_lines": {
                        "type": "integer"
                    },
                    "timeout": {
                        "type": "string"
                    },
                    "tui": {
                        "type": "boolean"
                    },
                    "verbose": {
                        "type": "boolean"
                    }
                },
                "type": "object"
            },
            "ConfigFetchReviews": {
                "properties": {
                    "nitpicks": {
                        "type": "boolean"
                    },
                    "provider": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "ConfigFixReviews": {
                "properties": {
                    "batch_size": {
                        "type": "integer"
                    },
                    "concurrent": {
                        "type": "integer"
                    },
                    "include_resolved": {
                        "type": "boolean"
                    },
                    "output_format": {
                        "type": "string"
                    },
                    "tui": {
                        "type": "boolean"
                    }
                },
                "type": "object"
            },
...
            "DashboardResponse": {
                "properties": {
                    "dashboard": {
                        "$ref": "#/components/schemas/DashboardPayload"
                    }
                },
                "required": ["dashboard"],
                "type": "object"
            },
...
            "ExtensionItem": {
                "properties": {
                    "description": {
                        "type": "string"
                    },
                    "enabled": {
                        "type": "boolean"
                    },
                    "name": {
                        "type": "string"
                    },
                    "source": {
                        "type": "string"
                    },
                    "version": {
                        "type": "string"
                    }
                },
                "required": ["name", "source", "enabled"],
                "type": "object"
            },
            "ExtensionListResponse": {
                "properties": {
                    "extensions": {
                        "items": {
                            "$ref": "#/components/schemas/ExtensionItem"
                        },
                        "type": "array"
                    },
                    "failure_count": {
                        "type": "integer"
                    },
                    "override_count": {
                        "type": "integer"
                    }
                },
                "required": ["extensions"],
                "type": "object"
            },
...
            "RunsResponse": {
                "properties": {
                    "runs": {
                        "items": {
                            "$ref": "#/components/schemas/Run"
                        },
                        "type": "array"
                    }
                },
                "required": ["runs"],
                "type": "object"
            },
...
            "SyncRequest": {
                "properties": {
                    "path": {
                        "type": "string"
                    },
                    "workflow_slug": {
                        "type": "string"
                    },
                    "workspace": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "SyncResult": {
                "properties": {
                    "checkpoints_updated": {
                        "type": "integer"
                    },
                    "legacy_artifacts_removed": {
                        "type": "integer"
                    },
                    "pruned_workflows": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array"
                    },
                    "review_issues_upserted": {
                        "type": "integer"
                    },
                    "review_rounds_upserted": {
                        "type": "integer"
                    },
                    "snapshots_upserted": {
                        "type": "integer"
                    },
                    "synced_at": {
                        "format": "date-time",
                        "type": "string"
                    },
                    "synced_paths": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array"
                    },
                    "target": {
                        "type": "string"
                    },
                    "task_items_upserted": {
                        "type": "integer"
                    },
                    "warnings": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array"
                    },
                    "workflow_slug": {
                        "type": "string"
                    },
                    "workflows_pruned": {
                        "type": "integer"
                    },
                    "workflows_scanned": {
                        "type": "integer"
                    },
                    "workspace_id": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "TransportError": {
                "properties": {
                    "code": {
                        "type": "string"
                    },
                    "details": {
                        "additionalProperties": true,
                        "type": "object"
                    },
                    "message": {
                        "type": "string"
                    },
                    "request_id": {
                        "type": "string"
                    }
                },
                "required": ["request_id", "code", "message"],
                "type": "object"
            },
...
            "WorkflowsResponse": {
                "properties": {
                    "workflows": {
                        "items": {
                            "$ref": "#/components/schemas/WorkflowSummary"
                        },
                        "type": "array"
                    }
                },
                "required": ["workflows"],
                "type": "object"
            },
...
        }
    },
    "info": {
        "title": "rc Daemon Browser API",
        "version": "0.1.0"
    },
    "openapi": "3.0.3",
    "paths": {
        "/api/catalog/agents": {
            "get": {
                "operationId": "listCatalogAgents",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/ActiveWorkspaceHeader"
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AgentListResponse"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "412": {
                        "$ref": "#/components/responses/StaleWorkspace"
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalError"
                    }
                },
                "summary": "List reusable agents available in the active workspace.",
                "tags": ["catalog"]
            }
        },
        "/api/catalog/extensions": {
            "get": {
                "operationId": "listCatalogExtensions",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/ActiveWorkspaceHeader"
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ExtensionListResponse"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "412": {
                        "$ref": "#/components/responses/StaleWorkspace"
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalError"
                    }
                },
                "summary": "List extensions installed in the active workspace.",
                "tags": ["catalog"]
            }
        },
...
        "/api/config/workspace": {
            "get": {
                "operationId": "getWorkspaceConfig",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/ActiveWorkspaceHeader"
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ConfigDocumentResponse"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "412": {
                        "$ref": "#/components/responses/StaleWorkspace"
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalError"
                    }
                },
                "summary": "Read the active workspace configuration.",
                "tags": ["config"]
            },
            "put": {
                "operationId": "putWorkspaceConfig",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/ActiveWorkspaceHeader"
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ConfigDocument"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ConfigDocumentResponse"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "400": {
                        "$ref": "#/components/responses/InvalidRequest"
                    },
                    "412": {
                        "$ref": "#/components/responses/StaleWorkspace"
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalError"
                    }
                },
                "summary": "Write the active workspace configuration.",
                "tags": ["config"]
            }
        },
...
        "/api/runs": {
            "get": {
                "operationId": "listRuns",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/WorkspaceFilterQuery"
                    },
                    {
                        "$ref": "#/components/parameters/RunStatus"
                    },
                    {
                        "$ref": "#/components/parameters/RunMode"
                    },
                    {
                        "$ref": "#/components/parameters/Limit"
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RunsResponse"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationError"
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalError"
                    }
                },
                "summary": "List runs across workspaces or for one workspace.",
                "tags": ["runs"]
            }
        },
...
        "/api/sync": {
            "post": {
                "operationId": "syncWorkflow",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/ActiveWorkspaceHeader"
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SyncRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SyncResult"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "400": {
                        "$ref": "#/components/responses/InvalidRequest"
                    },
                    "403": {
                        "$ref": "#/components/responses/Forbidden"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "412": {
                        "$ref": "#/components/responses/StaleWorkspace"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationError"
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalError"
                    }
                },
                "summary": "Run explicit daemon reconciliation for a workspace or workflow.",
                "tags": ["workflows"]
            }
        },
        "/api/tasks": {
            "get": {
                "operationId": "listWorkflows",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/ActiveWorkspaceHeader"
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/WorkflowsResponse"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "412": {
                        "$ref": "#/components/responses/StaleWorkspace"
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalError"
                    }
                },
                "summary": "List workflow summaries for the active workspace.",
                "tags": ["workflows"]
            }
        },
...
        "/api/ui/dashboard": {
            "get": {
                "operationId": "getDashboard",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/ActiveWorkspaceHeader"
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DashboardResponse"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "412": {
                        "$ref": "#/components/responses/StaleWorkspace"
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalError"
                    }
                },
                "summary": "Read the active-workspace dashboard aggregate payload.",
                "tags": ["dashboard"]
            }
        },
...
    },
...
}